Revert "Sell offers"
