--- .env.local
@@ -1,8 +1,0 @@
-NEXT_PUBLIC_PINATA_JWT='eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySW5mb3JtYXRpb24iOnsiaWQiOiJmNjZmOGJhMS05Mzk0LTQ5YjEtODI2MS1lYzZhZTExMDIwYjciLCJlbWFpbCI6ImxvdWlzLmNob3BpbmV0QGdtYWlsLmNvbSIsImVtYWlsX3ZlcmlmaWVkIjp0cnVlLCJwaW5fcG9saWN5Ijp7InJlZ2lvbnMiOlt7ImlkIjoiRlJBMSIsImRlc2lyZWRSZXBsaWNhdGlvbkNvdW50IjoxfSx7ImlkIjoiTllDMSIsImRlc2lyZWRSZXBsaWNhdGlvbkNvdW50IjoxfV0sInZlcnNpb24iOjF9LCJtZmFfZW5hYmxlZCI6ZmFsc2UsInN0YXR1cyI6IkFDVElWRSJ9LCJhdXRoZW50aWNhdGlvblR5cGUiOiJzY29wZWRLZXkiLCJzY29wZWRLZXlLZXkiOiIzNmQ2Nzk0NzU4YmQ2MWU4NzA5MSIsInNjb3BlZEtleVNlY3JldCI6IjdmYzM2YjE0MzhhNjEyNzQ5NzExNjU4MmY4MTdiZjZhMzY5ZjQ5ZjIxODUzMWQ0ODNmMDY0NzM0MzJkNTNhZmEiLCJpYXQiOjE3MTI0MDQ4MTl9.CmrugCfWb-n1_zkjqEnLypcLERe3o9w2v5KO5RuATtQ'
-NEXT_PUBLIC_GATEWAY_URL='https://harlequin-necessary-earthworm-821.mypinata.cloud'
-
-NEXT_PUBLIC_BROKER_ADRESS='rhMCswsPS5q7QuQUunWKPv9FqvW8ebkfw7'
-NEXT_PUBLIC_BROKER_SECRET='sEdViUZhM8Cbm8JuVg2EBmBJT141bXK'
-
-NEXT_PUBLIC_PROFILE_ADRESS='rQwTib8jtVgo3KVUannk2tYq6iaZ73Ky2K'
-NEXT_PUBLIC_PROFILE_SECRET='sEdTDxhkFNtcKuQsuViju4yNJgCpNBC'
