--- app/components/LoadWallet.jsx
@@ -1,11 +1,0 @@
-"use client";
-
-import { Wallet } from "@nice-xrpl/react-xrpl";
-
-export function LoadWallet({ children, seed }) {
-  return (
-    <Wallet seed={seed}>
-      {children}
-    </Wallet>
-  )
-}
