+++ app/components/navbar/index.tsx
@@ -69,9 +69,6 @@ const Navbar = ({ }: {
             <li>
               <Link href="/contact" onClick={toggleMobileNavbar} className="block py-2 pr-4 pl-3 hover:text-gray-400 lg:dark:hover:bg-transparent">contact</Link>
             </li>
-            <li>
-              <Link href='/profile' onClick={toggleMobileNavbar} className="block py-2 pr-4 pl-3 hover:text-gray-400 lg:dark:hover:bg-transparent">profile</Link>
-            </li>
           </ul>
         </div>
       </div>
