+++ package.json
@@ -15,7 +15,6 @@
     "react": "^18",
     "react-dom": "^18",
     "ripple-lib": "^1.10.1",
-    "xrpl": "^3.0.0",
     "xrpl-accountlib": "^3.2.9",
     "xrpl-client": "^2.3.1"
   },
